feat(Post 검색) : RDB를 통한 Post 검색 기능 구현

+++ seboard/src/main/java/com/seproject/seboard/domain/repository/comment/CommentSearchRepository.java
@@ -13,5 +13,5 @@ import java.util.List;
 public interface CommentSearchRepository extends Repository<Comment, Long> {
     Page<Comment> findCommentListByPostId(Long postId, Pageable pagingInfo);
     List<Reply> findReplyListByCommentId(Long commentId);
-    Long countReplyByPostId(Long postId);
+    int countReplyByPostId(Long postId);
 }
